requirements.txt
streamlit
langchain
langchain-community
langchain-groq
python-dotenv
ddgs
wikipedia